Set isPermaLink="no" for guids in rss feeds.
The rss spec says that unless the attribute is set, guid elements *are*
permalinks. The problem with that is that if [[meta permalink=]] is used,
as is done with aggregated posts, that goes into the link element, and
apparently some rss readers prefer the not-really-permalink in the guid
element when linking to the post.

Without meta permalink, the link and guid elements have the same content,
so it should be ok, in that case too for the guid to not be a permalink.

(Checked and this does not flood aggregators.)