revert massive l10n refersh
This reverts commit 5a82bc5e8799e2569cd256ef8274f00e8eff5d76
and commit 55dc24616892497a32e48ad349ebd16fac71e738.

That turns out to not be the way l10n of underlays is handled, instead
these changes are being made to the l10n branch, and the Danish
underlay translation update will be copied into master from that
branch.

This is necessary to avoid tying that branch up in a state where
it is horribly diverged from master and cannot be updated anymore.