new link to wmd zip file
wmd-editor.com is down, possibly forever. It's author seems to have pulled
a _why.

Ironically, the zip file is probably on netbsd's ftp server because
Schmonz packaged it for use for ikiwiki. In the event that it's removed
due to lack of license or something, I've added a copy to my personal
git-annex archive. Another source is
https://github.com/apenwarr/ekb/tree/master/wmd