add plugin tag type/comments and tag some pages.
I did this because I was looking for a summary/overview of the
various comment plugins that exist and couldn't already find
one.  I've tried to tag plugins that are mostly for use with
comments, which is a bit of a wide net and people may draw the
line in different places to myself.