add graphical explanations of various decentralisation schemes for ikiwiki
the wiki_edit_flow.svg file is slightly changed to make the box
designate the server boundaries so it can be reused in a coherent
manner in the other graphics.

the distributed_wikis page still needs an overhaul to clarify it's
about the third and fourth setups, but that will be easier to do in
separate commits.